requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiosignal==1.2.0
APScheduler==3.6.3
async-timeout==4.0.2
asynctest==0.13.0
atomicwrites==1.4.0
attrs==21.4.0
backtrader==1.9.76.123
beautifulsoup4==4.10.0
black==21.12b0
bs4==0.0.1
cachetools==4.2.2
ccxt==1.68.15
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.3
colorama==0.4.4
cycler==0.10.0
distlib==0.3.4
dnspython==2.1.0
docutils==0.18.1
fastquant==0.1.7.1
filelock==3.4.2
forex-python==1.8
frozenlist
networkx==2.6.3
nltk==3.6.7
nodeenv==1.6.0
numpy==1.21.2
oauthlib==3.1.1
packaging==21.3
pathspec==0.9.0
Pillow==8.3.2
platformdirs==2.4.1
pluggy==1.0.0
pre-commit==2.16.0
protobuf==3.20.1
py==1.11.0
pycares==4.1.2
pycoinmon==0.4.8
pycparser==2.21
Pygments==2.11.2
pymongo==3.12.0
pynput==1.7.6
pyparsing==2.4.7
PySimpleGUI==4.56.0
PySocks==1.7.1
pyspin==1.1.1
pytest==7.1.2
python-dateutil==2.8.2
python-telegram-bot==13.12
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
scipy==1.7.3
simplejson==3.17.6
six==1.16.0
soupsieve==2.2.1
tabulate==0.8.9
toml==0.10.2
tomli==1.2.3
tornado==6.1
tqdm==4.62.3
tweepy==4.4.0
typed-ast==1.5.1
typing_extensions==4.0.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.7
virtualenv==20.13.0
yarl==1.7.2
yfinance==0.1.69
zipp==3.7.0
